Factoring out RevisingResponseRenderer
